software/mac/requirements.txt
sounddevice>=0.4.6
webrtcvad>=2.0.10
numpy>=1.24.0
openai-whisper>=20231117
